[sledge] Evaluate function symbols applied to constants
Summary:
Function symbols when applied to literal values can be simplified by
evaluating them. This can be done even for function symbols that are
otherwise uninterpreted. This is not very strong, but is important in
some cases, and can prevent accumulating large complex terms that are
equal to literal constants.

Reviewed By: jvillard

Differential Revision: D24306044

fbshipit-source-id: 8c34d1ef2