[sledge] Refactor nonstdlib to avoid opening Core
Summary:
The treatment of comparison and exceptions in Core/Core_kernel/Base
makes them questionable as the default. This diff changes nonstdlib so
that Core is no longer opened in the global namespace, and makes a few
changes to handle the resulting minor API changes. This leads to a
lighter-touch nonstdlib, which makes a few definitions of its own, and
selects and extends modules from several libraries, including base,
core_kernel, containers, iter.

Reviewed By: jvillard

Differential Revision: D24306090

fbshipit-source-id: 42c91bd1b