[debug] new option --procedures-summary-json
Summary:
Output summaries in json format, so that other tools can exploit the
results of infer without having to be written inside infer itself.

For now the json for a summary is just one line saying "opaque" :)

Set up the infra to generate (yo)json automatically using
ppx_yojson_conv. See it in action in the next diff.

Reviewed By: ezgicicek

Differential Revision: D24503343

fbshipit-source-id: e24a2fff3