[preanalysis] Replace overridden method calls in ObjC
Summary:
This diff replaces overridden method calls in ObjC when possible, ie the first parameter of the
method has a sub-class type of the method's class.  For example,

when `MyEnumerator` is a sub-class of `NSEnumerator` and there is overridden `nextObject`,

```
[my_enumerator nextObject]
```

in Sil, it was translated to like

```
NSEnumerator.nextObject(my_enumerator : MyEnumerator*)
```

and the analyzer missed the overridden method.  This diff replaces the function call to

```
MyEnumerator.nextObject(my_enumerator : MyEnumerator*)
```

Reviewed By: ezgicicek

Differential Revision: D24477290

fbshipit-source-id: 6842a76f8