[frontend] Revise translation of message expression in ObjC frontend
Summary:
This diff revises the translation of message expression's arguments in ObjC frontend.  In the
frontend, it massages the arguments when calling a static method, so the class or object value is
not given to the static method as the first parameter.

The problem is that it used a raise-exception-and-catch way to detect where we remove the first
parameter.  This way of using an exception is not only hard to understand, but also incorrectly
removed the first parameter, with breaking abstract semantics sometimes.  (See the added test.)  This diff
avoids using the exception.

Reviewed By: jvillard

Differential Revision: D24565513

fbshipit-source-id: 0a84ca394