[nullsafe][refactor] Store all joinees' origins in InferredNullability
Summary:
The fact that we did not preserve all joins (and used only the first
one) was a shorcut since the origin is used only for reporting, and
reporting the info about left joinee is enough for practical purposes.

Hovewer, to support annotation graph, we need all joinees.

This diff is a no-op since we use only the first one currently.

Reviewed By: artempyanykh

Differential Revision: D24621412

fbshipit-source-id: 8fe1174e7