[pulse] Improve trace for Optional Empty Access
Summary: `folly::Optional::value()` returns a reference, hence an error was  shown when the actual value was being accessed. Since `value()` throws an exception in case of `folly::none`, we want to show the error message at the call site of `value()`. We do this by dereferencing the result of  `value()` in the model.

Reviewed By: jvillard

Differential Revision: D24702875

fbshipit-source-id: ca9f30349