[nullsafe][annotation graph] Support `Object.equals()`
Summary:
There is a feature in Nullsafe that is interfering with "annotation
graph" feature. Because of this we would not detect provisional
violations for misuses of params of equals() (They will be recorded
as user facing rather than provisional issues).

This diff turns this feature off for annotation graph mode.

Reviewed By: artempyanykh

Differential Revision: D24726655

fbshipit-source-id: 4b7577667