[nullsafe] Store field name in output json
Summary:
If the issue one of:
- Field Not Nullable
- Field Not Initialized
- Field Overannotated,

we record field_name to .json result.

NoTE: Design choice for representation. For Field Not Initialized and Field Overannotated
this is always internal (relative to the class) field, but for Field Not
Nullable it can be either internal or external. We could have a
structured output, or always output full name. I preferred to output
short name for convenience of the main usacase I am anticipating.

NOTE: not to be confused with the case where the field is nullable but
we e.g. try to dereference it. This is indirectly related to the issue
(can be several such fields for starters) and if we one day output it,
it will be provided in a separate way (similarly to how we output
nullable_methods).

Reviewed By: artempyanykh

Differential Revision: D24730320

fbshipit-source-id: c995ec221