[pulse] Less precise model for constructing optional from value
Summary: We recently introduced a more precise model for constructing an optional from a value by making a shallow copy. However, this introduced Use After Delete false positives. For now, we go back to a less precise model by creating a fresh value. A proper model would be to either make a deep copy or call the copy constructor for a value. We will address this in the following diff.

Reviewed By: jvillard

Differential Revision: D24826749

fbshipit-source-id: 3e5e4edeb