Fix of building Infer for Objective C (#1345)
Summary:
- The SDK path is different for Objective C when command line tools for Xcode are installed without Xcode itself.
- Related issue https://github.com/facebook/infer/issues/1002.

I still have a problem with the mentioned issue. I think that it is possible that the SDK paths differs for Objective C on Mac OS when I have installed just command line tools for Xcode without Xcode itself. This change solves the issue for me. Is it a good solution? Could it be possibly added to Infer? Or is there any better way?

Pull Request resolved: https://github.com/facebook/infer/pull/1345

Reviewed By: ngorogiannis

Differential Revision: D24917636

Pulled By: jvillard

fbshipit-source-id: 534ce4b49