[racerd] improve summary processing on calls
Summary: At a function call, an access performed by a callee must be processed in various ways before it's added to the accesses of the caller, and several of these steps may throw away the access.  Previously, this was done by effectively doing a bit of transformation, creating a new set of accesses, then folding over that to add to the caller's.  This is inefficient and somewhat confusing, as this can be done with one fold and a sequence of `Option.map`s.

Reviewed By: skcho

Differential Revision: D24885117

fbshipit-source-id: 4ab61eab9