[topl] Small steps in Pulse
Summary:
A Topl "small step" is a call to a method that is of interest to the
automaton. When such a call of interest is made, the topl component of
PulseAbductiveDomain.t is updated.  This means that intra-procedural
Topl should now work entirely inside Pulse, without instrumenting Sil.

Main TODOs:
 - add error extraction
 - implement inter-procedural (PulseTopl.large_step)

Reviewed By: jvillard

Differential Revision: D25028286

fbshipit-source-id: e31a96d13