[sledge] Fix a fresh name clash when solving extract equations
Summary:
The first-order solver sometimes needs to generate fresh variables to
express the solution of equations. It needs to ensure that these
generated variables do not clash. Before this diff, there was a
confusion where new variables were fresh with respect to only the
current set of "universal" variables. This is wrong, and this diff
adds the full set of variables instead.

Reviewed By: jvillard

Differential Revision: D25196732

fbshipit-source-id: afc56834a