[nullsafe] Fix behaviour of --no-nullsafe-optimistic-third-party-in-default-mode
Summary:
The problem is that in `AnnotatedField.special_case_nullability` we
first check the _generic_ nullability and if it is `nonnullish` we
apply refinements for enums, synthetic fields, etc.

The problem is that the definition of `is_nonnullish` changed in
D25186043 (https://github.com/facebook/infer/commit/7dcbacf69309d113e6374d03ee267fe2f7d44797) to a stricter one `UncheckedNonnull`, but generic
nullability stayed the same `ThirdPartyNonnull`.

Therefore enum elements were not considered `nonnullish` under
`--no-nullsafe-optimistic-third-party-in-default-mode` and the enum
refinements were not applied, which led to bogus errors.

**Example:**
There's a third-party enum
```
enum EnumClass {
  ENUM_ELEMENT
}
```

`ENUM_ELEMENT` is represented as a private static field of
`EnumClass`.

Then we have first party code that does
```
EnumClass.ENUM_ELEMENT
```

If this first party class is not `Nullsafe` and the checker is ran
with `--no-nullsafe-optimistic-third-party-in-default-mode`, the user
gets an incorrect warning about `ENUM_ELEMENT` being unvetted third
party.

Reviewed By: ngorogiannis

Differential Revision: D25560119

fbshipit-source-id: 4ad0760c5