[sledge] Rename Sh.is_false to is_unsat
Summary:
Sh.is_false checks if the pure constraints are inconsistent with the
first-order consequences of the spatial constraints. This involves
some, though not very expensive, logical reasoning. But is it not
checking only testing for the representation of Sh.false_, which one
might expect from its name. This renaming also makes room for the
operation that does test for the representation of Sh.false_.

Reviewed By: jvillard

Differential Revision: D25756580

fbshipit-source-id: 30510f45a